requirements.txt
fastapi
uvicorn
python-dotenv
requests
anthropic
faiss-cpu
numpy
pydantic